fix: update test
